PYIC-8593: Extract nested transition traffic handler to separate file.
